+++ routes/web.php
@@ -1,7 +1,7 @@
 <?php
 
 use App\Http\Controllers\CommentController;
-use App\Http\Controllers\LikeController;
+use App\Http\Controllers\LikeToggleController;
 use App\Http\Controllers\PostController;
 use Illuminate\Support\Facades\Route;
 use Inertia\Inertia;
@@ -28,5 +28,4 @@ Route::post('/posts', [PostController::class, 'store']);
 
 Route::post('/comments', [CommentController::class, 'store']);
 
-Route::post('/posts/{post}/like', [LikeController::class, 'store']);
-Route::delete('/posts/{post}/like', [LikeController::class, 'destroy']);
+Route::post('/posts/{post}/like/toggle', [LikeToggleController::class, 'store']);
